Ajoute le bouton Accéder aux résultats lorsque la simulation est terminée
